requirements.txt
# API Framework
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
pydantic-settings>=2.0.0

# RAG Components
numpy>=1.24.0
faiss-cpu>=1.7.4
tiktoken>=0.5.0
langchain>=0.0.267
langchain-openai>=0.0.1
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6